refactor: move table config
