+++ src/apps/ScopeChangeRequest/utils/config/index.ts
@@ -4,3 +4,4 @@ export * from './dataOptions';
 export * from './prefetchQueryOptions';
 export * from './filter/filterConfig';
 export * from './garden/gardenConfig';
+export * from './table';
