+++ src/apps/ScopeChangeRequest/utils/config/table/index.ts
@@ -1,0 +1,1 @@
+export * from './tableConfig';
